requirement.txt
amqp==2.5.2
APScheduler==3.6.3
asgiref==3.2.3
attrs==19.3.0
autobahn==20.1.2
Automat==0.8.0
bcrypt==3.1.7
billiard==3.6.3.0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
comtypes==1.1.7
constantly==15.1.0
cryptography==2.8
daphne==2.4.1
diff-match-patch==20181111
Django==2.1.7
django-apscheduler==0.3.0
django-crontab==0.7.1
django-debug-toolbar==2.2
django-import-export==1.2.0
django-middleware-global-request==0.2.0
django-timezone-field==4.0
docutils==0.16
docxtpl==0.6.3
dwebsocket==0.5.12
fpdf2==2.0.3
future==0.18.2
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.5.0
incremental==17.5.0
Jinja2==2.10.3
kombu==4.6.7
ldap3==2.8.1
lxml==4.4.2
MarkupSafe==1.1.1
numpy==1.18.0
pandas==0.25.3
paramiko==2.7.1
Pillow==7.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
PyHamcrest==2.0.0
PyMySQL==0.9.3
PyNaCl==1.3.0
pyOpenSSL==19.1.0
PyPDF2==1.26.0
PyPDF4==1.27.0
python-crontab==2.4.0
python-dateutil==2.8.1
python-docx==0.8.7
pytz==2019.3
pywin32==227
reportlab==3.5.34
requests==2.22.0
retrying==1.3.3
service-identity==18.1.0
six==1.13.0
spyne==2.13.16
SQLAlchemy==1.3.13
sqlparse==0.3.0
suds-community==0.8.4
tablib==1.1.0
Twisted==19.10.0
txaio==18.8.1
tzlocal==2.0.0
urllib3==1.25.7
vine==1.3.0
xlrd==1.2.0
xlwt==1.3.0
zipp==3.0.0
zope.interface==4.7.1
